requirements.txt
cffi==1.17.1
mypy==1.14.1
mypy-extensions==1.0.0
pycparser==2.22
PyNaCl==1.5.0
typing_extensions==4.12.2
